Rewrote the NotificationController to use the StreakController
